TNE-MP README.md
TNE_SOCIETY Music Player Prototype
Note: This is a Prototype

This version of the TNE_SOCIETY Music Player is a prototype built on localhost for testing and development purposes. The final project will be developed on the BOS (Blockchain Operating System) to leverage the power of blockchain technology within the NEAR ecosystem.